move generate certificate to single package
